Changed folder structure
