--- app/(routes)/admin/AdminPage.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import AppLayout from "@/app/_common/layout/AppLayout";
-import AdminContent from "./AdminContent";
-
-const AdminPage = () => {
-  return (
-    <AppLayout>
-      <AdminContent />
-    </AppLayout>
-  );
-};
-
-export default AdminPage;
