--- app/(routes)/admin/component-wall/ComponentWallPage.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import AppLayout from "@/app/_common/layout/AppLayout";
-import ComponentWallContent from "./ComponentWallContent";
-
-const ComponentWallPage = () => {
-  return (
-    <AppLayout>
-      <ComponentWallContent />
-    </AppLayout>
-  );
-};
-
-export default ComponentWallPage;
