+++ app/(routes)/dashboard/_components/DashboardPage.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+const DashboardPage = () => {
+  return (
+    <div className="text-4xl">
+      This is the dashboard page.
+    </div>
+  );
+};
+
+export default DashboardPage;
